slack importer: Move message conversions to a new module.
